requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<2.0
requests>=2.31,<3.0
alpaca-py>=0.21,<1.0
xgboost>=2.0,<3.0
scikit-learn>=1.3,<2.0
python-dotenv>=1.0,<2.0
ta>=0.11,<1.0
joblib>=1.3,<2.0
openai>=1.0,<2.0
pytz>=2023.3
